Moved all dialogs to separate files.
